+++ packages/milkup-youtube/src/index.tsx
@@ -1,45 +1,2 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { $insertNodeToNearestRoot } from "@lexical/utils";
-import {
-  COMMAND_PRIORITY_EDITOR,
-  createCommand,
-  LexicalCommand,
-} from "lexical";
-import { useEffect } from "react";
-
-import { $createYouTubeNode, YouTubeNode } from "./YoutubeNode";
-
-export const INSERT_YOUTUBE_COMMAND: LexicalCommand<string> = createCommand(
-  "INSERT_YOUTUBE_COMMAND",
-);
-
-export default function YouTubePlugin(): JSX.Element | null {
-  const [editor] = useLexicalComposerContext();
-
-  useEffect(() => {
-    if (!editor.hasNodes([YouTubeNode])) {
-      throw new Error("YouTubePlugin: YouTubeNode not registered on editor");
-    }
-
-    return editor.registerCommand<string>(
-      INSERT_YOUTUBE_COMMAND,
-      (payload) => {
-        const youTubeNode = $createYouTubeNode(payload);
-        $insertNodeToNearestRoot(youTubeNode);
-
-        return true;
-      },
-      COMMAND_PRIORITY_EDITOR,
-    );
-  }, [editor]);
-
-  return null;
-}
+export * from "./YoutubeNode";
+export * from "./YoutubePlugin";
